+++ src/main/java/com/github/jdavies100/debts/service/DebtService.java
@@ -1,12 +1,13 @@
 package com.github.jdavies100.debts.service;
 
 import com.github.jdavies100.debts.model.Person;
+import java.math.BigDecimal;
 import org.springframework.stereotype.Component;
 
 @Component
 public interface DebtService {
 
-  Person addDebt(String personId, double amount) throws Exception;
+  Person addDebt(String personId, BigDecimal amount) throws Exception;
 
   Person removeDebt(String personId, String debtId) throws Exception;
